refactor: 테스트 픽스쳐 위치 변경
